Fix UTs

+++ pkg/workflow/argo/fetch_test.go
@@ -215,9 +215,9 @@ func TestFetch(t *testing.T) {
 	a.Contains(concatedLogs, "SQLFlow Step: [3/3] Status: Succeeded")
 	// confirm columns and rows of sql: SELECT 1;
 	a.Equal([]string{"1"}, columns)
-	v := &wrappers.Int64Value{}
+	v := &wrappers.DoubleValue{} // FIXME(shendiaomo): json doesn't have numeric type
 	a.NoError(ptypes.UnmarshalAny(rows[0][0], v))
-	a.Equal(v.GetValue(), int64(1))
+	a.Equal(v.GetValue(), float64(1))
 }
 
 func waitUntilPodRunning(podID string) error {
